Use android service for long running forward tasks
