+++ .gitignore
@@ -15,3 +15,4 @@ build/
 .cxx
 local.properties
 venv/
+yarn.lock
