+++ androidApp/src/main/AndroidManifest.xml
@@ -29,6 +29,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <service android:name=".MeasurementService"
+            android:foregroundServiceType="location|microphone"
+            />
     </application>
 
 </manifest>
