+++ gradle/libs.versions.toml
@@ -12,14 +12,14 @@ androidx-test-espresso-core = "3.5.1"
 androidx-test-junit = "1.1.5"
 appyx = "2.0.0-alpha09"
 compose-ui-test-junit4 = "1.6.2"
-compose-compiler = "1.6.1"
+compose-compiler = "1.5.11"
 compose-plugin = "1.6.1"
 core-ktx = "1.12.0"
 coroutines = "1.8.0"
 junit = "4.13.2"
 jvm-target = "1.8"
 koin-android = "3.4.0"
-kotlin = "1.9.22"
+kotlin = "1.9.23"
 ksp = "1.9.22-1.0.17"
 koin-core = "3.4.0"
 appcompat = "1.6.1"
